pyproject.toml
[tool.poetry]
name = "export"
version = "0.1.0"
description = ""
authors = ["Saeed Gholami <saeed.ghollami@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
pillow = "^10.0.0"
django-crispy-forms = "^2.0"
crispy-bootstrap5 = "^0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
